requirements.txt
flask
transformers
pillow
requests
torch